requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
insightface>=0.7.3
onnxruntime>=1.16.0
numpy
sqlalchemy
pillow
python-multipart
